add proper PackerConfig object, add tests!

+++ src/modpacker/server.py
@@ -9,18 +9,18 @@ import tqdm
 from tqdm.contrib.logging import logging_redirect_tqdm
 
 from modpacker.compile import read_or_download, unsup_ini_content
-from modpacker.config import open_config
+from modpacker.packer_config import PackerConfig
 
 logger = logging.getLogger(__name__)
 
 
-def export(output_folder):
-    packer_config = open_config()
+def export(packer_config: PackerConfig, output_folder, unattended = False):
     output_folder = os.path.realpath(output_folder)
     if os.path.exists(output_folder):
-        answer = questionary.confirm(f"This will delete {os.path.realpath(output_folder)}! Do you want to continue?").ask()
-        if not answer:
-            return
+        if not unattended:
+            answer = questionary.confirm(f"This will delete {os.path.realpath(output_folder)}! Do you want to continue?").ask()
+            if not answer:
+                return
         shutil.rmtree(output_folder)
     os.makedirs(output_folder)
 
@@ -46,7 +46,7 @@ def export(output_folder):
                     with open(os.path.join(output_folder, path), "wb") as f:
                         f.write(data)
 
-    if "dependencies" in packer_config and "neoforge" in packer_config["dependencies"]:
+    if packer_config.mod_loader == "neoforge":
         logger.info("Neoforge detected. Downloading installer...")
         neoforge_version = packer_config['dependencies']['neoforge']
         neoforge_installer = f"https://maven.neoforged.net/releases/net/neoforged/neoforge/{neoforge_version}/neoforge-{neoforge_version}-installer.jar"
@@ -55,19 +55,19 @@ def export(output_folder):
             f.write(installer_data.content)
         logger.info(f"Downloaded Neoforge installer in {os.path.join(output_folder, 'neo-installer.jar')}.")
 
-        answer = questionary.confirm("Run server installer?").ask()
-        if answer:
-            installer_process = subprocess.Popen(["java", "-jar", os.path.join(output_folder, "neo-installer.jar"), "--install-server"], cwd=output_folder)
-            code = installer_process.wait()
-            if code == 0:
-                answer = questionary.confirm("Delete Neoforge installer?").ask()
-                if answer:
-                    os.remove(os.path.join(output_folder, "neo-installer.jar"))
-            else:
-                logger.error("Neoforge installer encountered an error.")
+        if not unattended:
+            answer = questionary.confirm("Run server installer?").ask()
+            if answer:
+                installer_process = subprocess.Popen(["java", "-jar", os.path.join(output_folder, "neo-installer.jar"), "--install-server"], cwd=output_folder)
+                code = installer_process.wait()
+                if code == 0:
+                    answer = questionary.confirm("Delete Neoforge installer?").ask()
+                    if answer:
+                        os.remove(os.path.join(output_folder, "neo-installer.jar"))
+                else:
+                    logger.error("Neoforge installer encountered an error.")
 
-
-    if "unsup" in packer_config:
+    if packer_config.has_unsup():
         logger.info("Downloading unsup.jar and creating unsup.ini...")
         latest_release = requests.get(f"https://git.sleeping.town/api/v1/repos/exa/unsup/releases/tags/v{packer_config['unsup']['version']}").json()
         for asset in latest_release['assets']:
